requirements.txt
flask==2.0.1
werkzeug==2.0.1
python-dotenv==0.19.1
pandas==1.3.3
PyPDF2==1.26.0
python-docx==0.8.11
openai==1.3.0
httpx==0.24.1
numpy==1.21.2
requests==2.26.0